more tests in new class
