+++ src/test/java/software/amazon/encryption/s3/S3EncryptionClientTest.java
@@ -1102,81 +1102,6 @@ public class S3EncryptionClientTest {
         s3Client.close();
     }
 
-    @Test
-    public void testPutWithInstructionFile() {
-        final String objectKey = appendTestSuffix("instruction-file-put-object");
-        final String objectKeyV2 = appendTestSuffix("instruction-file-put-object-v2");
-        final String input = "SimpleTestOfV3EncryptionClient";
-        S3Client wrappedClient = S3Client.create();
-        S3Client s3Client = S3EncryptionClient.builder()
-                .instructionFileConfig(InstructionFileConfig.builder()
-                        .instructionFileClient(wrappedClient)
-                        .enableInstructionFilePutObject(true)
-                        .build())
-                .kmsKeyId(KMS_KEY_ID)
-                .build();
-
-        s3Client.putObject(builder -> builder
-                .bucket(BUCKET)
-                .key(objectKey)
-                .build(), RequestBody.fromString(input));
-
-        // Disabled client should fail
-        S3Client s3ClientDisabledInstructionFile = S3EncryptionClient.builder()
-                .wrappedClient(wrappedClient)
-                .instructionFileConfig(InstructionFileConfig.builder()
-                        .disableInstructionFile(true)
-                        .build())
-                .kmsKeyId(KMS_KEY_ID)
-                .build();
-
-        try {
-            s3ClientDisabledInstructionFile.getObjectAsBytes(builder -> builder
-                    .bucket(BUCKET)
-                    .key(objectKey)
-                    .build());
-            fail("expected exception");
-        } catch (S3EncryptionClientException exception) {
-            assertTrue(exception.getMessage().contains("Exception encountered while fetching Instruction File."));
-        }
-
-        // Get the instruction file separately using a default client
-        S3Client defaultClient = S3Client.create();
-        ResponseBytes<GetObjectResponse> directInstGetResponse = defaultClient.getObjectAsBytes(builder -> builder
-                .bucket(BUCKET)
-                .key(objectKey + ".instruction")
-                .build());
-        assertTrue(directInstGetResponse.response().metadata().containsKey("x-amz-crypto-instr-file"));
-
-        ResponseBytes<GetObjectResponse> objectResponse = s3Client.getObjectAsBytes(builder -> builder
-                .bucket(BUCKET)
-                .key(objectKey)
-                .build());
-        String output = objectResponse.asUtf8String();
-        assertEquals(input, output);
-
-
-        // Temporary - Generate an instruction file in V2 to compare against V3
-        // TODO: do this for other keyrings as well
-        EncryptionMaterialsProvider materialsProvider =
-                new StaticEncryptionMaterialsProvider(new KMSEncryptionMaterials(KMS_KEY_ID));
-        CryptoConfigurationV2 cryptoConfig =
-                new CryptoConfigurationV2(CryptoMode.StrictAuthenticatedEncryption)
-                        .withStorageMode(CryptoStorageMode.InstructionFile);
-
-        AmazonS3EncryptionV2 v2Client = AmazonS3EncryptionClientV2.encryptionBuilder()
-                .withCryptoConfiguration(cryptoConfig)
-                .withEncryptionMaterialsProvider(materialsProvider)
-                .build();
-
-        v2Client.putObject(BUCKET, objectKeyV2, input);
-
-        // Cleanup
-//        deleteObject(BUCKET, objectKey, s3Client);
-        s3ClientDisabledInstructionFile.close();
-        s3Client.close();
-    }
-
     /**
      * A simple, reusable round-trip (encryption + decryption) using a given
      * S3Client. Useful for testing client configuration.
